XFree86 4.3.0.1

+++ src/TMprint.c
@@ -58,6 +58,7 @@ used in advertising or otherwise to promote the sale, use or other dealings
 in this Software without prior written authorization from The Open Group.
 
 */
+/* $XFree86: xc/lib/Xt/TMprint.c,v 1.6 2001/12/14 19:56:31 dawes Exp $ */
 
 /*LINTLIBRARY*/
 #include "IntrinsicI.h"
@@ -196,7 +197,7 @@ static void PrintCode(sb, mask, code)
     if (mask != 0) {
 	if (mask != (unsigned long)~0L)
 	    (void) sprintf(sb->current, "0x%lx:0x%lx", mask, code);
-	else (void) sprintf(sb->current, /*"0x%lx"*/ "%d", code);
+	else (void) sprintf(sb->current, /*"0x%lx"*/ "%d", (unsigned)code);
 	sb->current += strlen(sb->current);
     }
 }
